# Optimizing math functions like exp, log and trigonometric