1. Exponent: Optimized using binary exponentiation for large values. For exponent(1000), it gave an improvement in speed by taking 0.8s now as compared to 256s earlier. Also improved in cases where x<-1.
2. Logarithm: Wrote log(x)=log(x/2)+log(2), repeating until x is small enough. Also calculated log(2) only once. For log(100), it gave an improvement in speed by taking 5s now as compared to 437s earlier. Also improved in cases where x<1.
3. Trigonometric Functions: Performed argument reduction to reduce the number to [0,2PI). Then used Taylor series. For sin(10000), it gave an improvement in speed by taking 0.6s now as compared to 64s earlier. Implemented floor function for it.
4. Added sqrt and inverse trigonometric functions.

Also using up_to() at each iteration of taylor series to avoid extra computation. This also gave significant improvement in time.