@div5252 2021-07-15
@vikram2000b As I have mentioned in issue #5, the problem with taylor expansion for large inputs of trigonometric function is that they take many iterations before converging.
eg. sin(1000) took 1400 iterations. And there is also another issue that sin(1000) gives incorrect result (not b/w -1 and 1 as it should be) due to less precision after so many iterations.
The pi algorithm isn't slow as the complexity of Chudnovsky algorithm is O(n(logn)^3). And for our library 10^9 is still a small number, so it doesn't require many digits of pi.
Anyways, I also calculated the performances as mentioned in the description of PR-
"For sin(10000), it gave an improvement in speed by taking 0.6s now as compared to 64s earlier." And even for large inputs, my modified sine function takes less than a second. I'll demonstrate in the call.