@Laouen 2021-07-15
I agree with @vikram2000b about adding more comments to explain the code calculations and improve readability. Regarding the optimizations, I think that even though you tried with an example and you found significant improvements, you can be doing cherry-picking in a case where it improves but it can occur (not saying that will occur) that in general, it is worse. Suraj is working with benchmarking and it would be great to be able to compare original implementation vs improvements.